+++ hspec-core/src/Test/Hspec/Core/Runner/Util.hs
@@ -1,5 +1,12 @@
 {-# LANGUAGE CPP #-}
-module Test.Hspec.Core.Runner.Util where
+module Test.Hspec.Core.Runner.Util (
+  extractLocation
+
+-- for testing
+, parseCallStack
+, parseLocation
+, parseSourceSpan
+) where
 
 import           Prelude ()
 import           Test.Hspec.Core.Compat
@@ -7,16 +14,25 @@ import           Test.Hspec.Core.Compat
 import           Control.Exception
 import           Data.List
 import           Data.Char
+import           Data.Maybe
 
 import           Test.Hspec.Core.Example
 
 extractLocation :: SomeException -> Maybe Location
-extractLocation e = case fromException e of
+extractLocation e = locationFromErrorCall e <|> locationFromPatternMatchFail e
+
+locationFromErrorCall :: SomeException -> Maybe Location
+locationFromErrorCall e = case fromException e of
 #if MIN_VERSION_base(4,9,0)
   Just (ErrorCallWithLocation _ loc) -> parseCallStack loc
 #else
   Just (ErrorCall _) -> Nothing
 #endif
+  Nothing -> Nothing
+
+locationFromPatternMatchFail :: SomeException -> Maybe Location
+locationFromPatternMatchFail e = case fromException e of
+  Just (PatternMatchFail s) -> listToMaybe (words s) >>= parseSourceSpan
   Nothing -> Nothing
 
 parseCallStack :: String -> Maybe Location
@@ -34,5 +50,20 @@ parseCallStack input = case reverse (lines input) of
 parseLocation :: String -> Maybe Location
 parseLocation input = case fmap breakColon (breakColon input) of
   (file, (line, column)) -> Location file <$> readMaybe line <*> readMaybe column
-  where
-    breakColon = fmap (drop 1) . break (== ':')
+
+parseSourceSpan :: String -> Maybe Location
+parseSourceSpan input = case breakColon input of
+  (file, xs) -> (uncurry $ Location file) <$> (tuple <|> colonSeparated)
+    where
+      lineAndColumn :: String
+      lineAndColumn = takeWhile (/= '-') xs
+
+      tuple :: Maybe (Int, Int)
+      tuple = readMaybe lineAndColumn
+
+      colonSeparated :: Maybe (Int, Int)
+      colonSeparated = case breakColon lineAndColumn of
+        (l, c) -> (,) <$> readMaybe l <*> readMaybe c
+
+breakColon :: String -> (String, String)
+breakColon = fmap (drop 1) . break (== ':')