+++ hspec-core/test/Test/Hspec/Core/Runner/UtilSpec.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE CPP #-}
+{-# OPTIONS_GHC -fno-warn-incomplete-patterns #-}
 module Test.Hspec.Core.Runner.UtilSpec (spec) where
 
 import           Helper
@@ -10,16 +11,33 @@ import           Test.Hspec.Core.Runner.Util
 spec :: Spec
 spec = do
   describe "extractLocation" $ do
-    it "extracts Location from exception" $ do
-      let
-        location =
+    context "with ErrorCall" $ do
+      it "extracts Location" $ do
+        let
+          location =
 #if MIN_VERSION_base(4,9,0)
-          Just $ Location __FILE__ (__LINE__ + 4) 32
+            Just $ Location __FILE__ (__LINE__ + 4) 34
 #else
-          Nothing
+            Nothing
 #endif
-      Left e <- try (evaluate (undefined :: ()))
-      extractLocation e `shouldBe` location
+        Left e <- try (evaluate (undefined :: ()))
+        extractLocation e `shouldBe` location
+
+    context "with PatternMatchFail" $ do
+      context "with single-line source space" $ do
+        it "extracts Location" $ do
+          let
+            location = Just $ Location __FILE__ (__LINE__ + 1) 40
+          Left e <- try (evaluate (let Just n = Nothing in (n :: Int)))
+          extractLocation e `shouldBe` location
+
+      context "with multi-line source space" $ do
+        it "extracts Location" $ do
+          let location = Just $ Location __FILE__ (__LINE__ + 1) 36
+          Left e <- try (evaluate (case Nothing of
+            Just n -> n :: Int
+            ))
+          extractLocation e `shouldBe` location
 
   describe "parseCallStack" $ do
     it "parses Location from call stack" $ do
@@ -33,3 +51,10 @@ spec = do
   describe "parseLocation" $ do
     it "parses Location" $ do
       parseLocation "test/Test/Hspec.hs:13:32" `shouldBe` Just (Location "test/Test/Hspec.hs" 13 32)
+
+  describe "parseSourceSpan" $ do
+    it "parses single-line source span" $ do
+      parseSourceSpan "test/Test/Hspec.hs:25:36-51:" `shouldBe` Just (Location "test/Test/Hspec.hs" 25 36)
+
+    it "parses multi-line source span" $ do
+      parseSourceSpan "test/Test/Hspec.hs:(15,7)-(17,26):" `shouldBe` Just (Location "test/Test/Hspec.hs" 15 7)